+++ RELEASE_NOTES.md
@@ -5,5 +5,6 @@
 ## New Features
 
 - Added HMAC capabilities
+- Add option to request public order data from the API.
 
 ## Bug Fixes